cambio

+++ src/main/java/com/intech/ptmsusuarios/dao/IUserDao.java
@@ -5,4 +5,4 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IUserDao extends CrudRepository<User, Integer> { }
+public interface IUserDao extends CrudRepository<User, Long> { }
